example/otlp-traces/requirements.txt
opentelemetry-sdk==1.17.0
opentelemetry-exporter-otlp==1.17.0
opentelemetry-sdk-extension-aws==2.0.1
